Use common metric upload code
