+++ packages/webapp/src/routes/developer/check-metric-file.js
@@ -36,7 +36,7 @@ const handlers = {
 const getContext = request => {
   const fileLocation = request.yar.get(constants.redisKeys.DEVELOPER_METRIC_LOCATION)
   const fileSize = request.yar.get(constants.redisKeys.DEVELOPER_METRIC_FILE_SIZE)
-  const humanReadableFileSize = getHumanReadableFileSize(fileSize)
+  const humanReadableFileSize = getHumanReadableFileSize(fileSize, 1)
   return {
     filename: fileLocation === null ? '' : path.parse(fileLocation).base,
     fileSize: humanReadableFileSize
