+++ packages/webapp/src/utils/build-metric-upload-config.js
@@ -1,0 +1,51 @@
+import constants from './constants.js'
+import { uploadStreamAndQueueMessage } from './azure-storage.js'
+import { handleEvents } from './azure-signalr.js'
+
+const buildConfig = (sessionId, metricUploadType) => {
+  const config = { metricUploadType }
+  buildBlobConfig(sessionId, config)
+  buildQueueConfig(config)
+  buildFunctionConfig(config)
+  buildSignalRConfig(sessionId, config)
+  buildFileValidationConfig(config)
+  return config
+}
+
+const buildBlobConfig = (sessionId, config) => {
+  config.blobConfig = {
+    blobName: `${sessionId}/${config.metricUploadType}/`,
+    containerName: 'untrusted'
+  }
+}
+
+const buildQueueConfig = config => {
+  config.queueConfig = {
+    uploadType: config.metricUploadType,
+    queueName: 'untrusted-file-queue'
+  }
+}
+
+const buildFunctionConfig = config => {
+  config.functionConfig = {
+    uploadFunction: uploadStreamAndQueueMessage,
+    handleEventsFunction: handleEvents
+  }
+}
+
+const buildSignalRConfig = (sessionId, config) => {
+  config.signalRConfig = {
+    eventProcessingFunction: null,
+    timeout: parseInt(process.env.UPLOAD_PROCESSING_TIMEOUT_MILLIS) || 180000,
+    url: `${process.env.SIGNALR_URL}?userId=${sessionId}`
+  }
+}
+
+const buildFileValidationConfig = config => {
+  config.fileValidationConfig = {
+    fileExt: constants.metricFileExt,
+    maxFileSize: parseInt(process.env.MAX_METRIC_UPLOAD_MB) * 1024 * 1024
+  }
+}
+
+export { buildConfig }
